feature
1. 레시피 상세페이지 작성
2. 데이터통신 코드 viewController로 옮겨야 함

+++ PickyCookBook/PickyCookBook/Recipe.swift
@@ -15,7 +15,7 @@ struct Recipe {
     var img_recipe: String
     var ingredient: String
     let description: String
-    var like_count: Double
+    var like_count: Int
     var rate_sum: Double
     var tag: String
